feat(plugin-notification-manager): support handlebars in message
