+++ packages/core/utils/package.json
@@ -11,6 +11,7 @@
     "deepmerge": "^4.2.2",
     "flat-to-nested": "^1.1.1",
     "graphlib": "^2.1.8",
+    "handlebars": "^4.7.8",
     "multer": "^1.4.5-lts.1",
     "object-path": "^0.11.8"
   },
