+++ packages/core/utils/src/index.ts
@@ -17,6 +17,7 @@ export * from './date';
 export * from './dayjs';
 export * from './forEach';
 export * from './fs-exists';
+export * from './handlebars';
 export * from './isValidFilter';
 export * from './json-templates';
 export * from './koa-multer';
